assessment-takehome/takehome-fairmoney/requirements.txt
altair==4.2.2
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
deap==1.3.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
flit_core==3.8.0
fonttools==4.25.0
gitdb==4.0.10
GitPython==3.1.31
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.10.1
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.6
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingpy==0.2.0
mistune==0.8.4
multimethod==1.9.1
munkres==1.1.4
nbclassic==0.5.4
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
plotly==5.14.1
ply==3.11
pooch==1.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.4.2
QtPy==2.2.0
requests==2.28.1
rich==13.3.4
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
shap==0.41.0
sip==6.6.2
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
statsmodels==0.13.5
stopit==1.1.2
streamlit==1.21.0
sweetviz==2.1.4
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==2.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
TPOT==0.11.7
tqdm==4.64.1
traitlets==5.7.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
update-checker==0.18.0
urllib3==1.26.15
validators==0.20.0
visions==0.7.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
widgetsnbextension==4.0.5
xgboost==1.7.5
ydata-profiling==4.1.2
zipp==3.11.0
